working userspace C printf
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@32 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ lib/printfmt.c
@@ -424,7 +424,7 @@ sprintputch (int ch, struct sprintbuf *b)
 }
 
 int
-vsnprintf (char *buf, int n, const char *fmt, va_list ap)
+vsnprintf (char *buf, size_t n, const char *fmt, va_list ap)
 {
   struct sprintbuf b = { buf, buf + n - 1, 0 };
 
@@ -441,7 +441,7 @@ vsnprintf (char *buf, int n, const char *fmt, va_list ap)
 }
 
 int
-snprintf (char *buf, int n, const char *fmt, ...)
+snprintf (char *buf, size_t n, const char *fmt, ...)
 {
   va_list ap;
   int rc;
